requirements.txt
djangorestframework
Django==4.2.7
requests